simple refactoring
